src/main/webapp/app/entities/probleme/detail/probleme-detail.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { ActivatedRoute } from '@angular/router';
import { of } from 'rxjs';

import { DataUtils } from 'app/core/util/data-util.service';

import { ProblemeDetailComponent } from './probleme-detail.component';

describe('Component Tests', () => {
  describe('Probleme Management Detail Component', () => {
    let comp: ProblemeDetailComponent;
    let fixture: ComponentFixture<ProblemeDetailComponent>;
    let dataUtils: DataUtils;

    beforeEach(() => {
      TestBed.configureTestingModule({
        declarations: [ProblemeDetailComponent],
        providers: [
          {
            provide: ActivatedRoute,
            useValue: { data: of({ probleme: { id: 123 } }) },
          },
        ],
      })
        .overrideTemplate(ProblemeDetailComponent, '')
        .compileComponents();
      fixture = TestBed.createComponent(ProblemeDetailComponent);
      comp = fixture.componentInstance;
      dataUtils = TestBed.inject(DataUtils);
      jest.spyOn(window, 'open').mockImplementation(() => null);
    });

    describe('OnInit', () => {
      it('Should load probleme on init', () => {
        // WHEN
        comp.ngOnInit();

        // THEN
        expect(comp.probleme).toEqual(expect.objectContaining({ id: 123 }));
      });
    });

    describe('byteSize', () => {
      it('Should call byteSize from DataUtils', () => {
        // GIVEN
        jest.spyOn(dataUtils, 'byteSize');
        const fakeBase64 = 'fake base64';

        // WHEN
        comp.byteSize(fakeBase64);

        // THEN
        expect(dataUtils.byteSize).toBeCalledWith(fakeBase64);
      });
    });

    describe('openFile', () => {
      it('Should call openFile from DataUtils', () => {
        // GIVEN
        jest.spyOn(dataUtils, 'openFile');
        const fakeContentType = 'fake content type';
        const fakeBase64 = 'fake base64';

        // WHEN
        comp.openFile(fakeBase64, fakeContentType);

        // THEN
        expect(dataUtils.openFile).toBeCalledWith(fakeBase64, fakeContentType);
      });
    });
  });
});
